Periodo 2017.2/Programação Para Internet/geoufpi/requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.0
cairocffi==0.8.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.4
cssselect==1.0.1
cssselect2==0.2.0
dj-database-url==0.4.2
Django==1.11
django-appconf==1.0.2
django-decouple==2.1
django-filter==1.1.0
django-material==1.1.1
django-viewflow==1.1.0
djangorestframework==3.7.3
gunicorn==19.7.1
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
lxml==4.1.1
Markdown==2.6.10
olefile==0.44
pilkit==2.0
Pillow==4.3.0
pipdeptree==0.10.1
psycopg2==2.7.3.2
pycparser==2.18
pylokit==0.8.1
pyOpenSSL==17.5.0
PyPDF2==1.26.0
Pyphen==0.9.4
pytz==2017.3
reportlab==3.4.0
requests==2.18.4
semantic-version==2.6.0
six==1.11.0
templated-docs==0.3.1
tinycss==0.4
tinycss2==0.6.0
updatable==0.2.0
urllib3==1.22
webencodings==0.5.1
whitenoise==3.3.1
xhtml2pdf==0.2b1
